requirements.txt
art==4.6
astroid==2.4.1
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
libmagic==1.0
mccabe==0.6.1
numpy==1.18.4
Pillow==7.1.2
pylint==2.5.2
python-magic==0.4.18
requests==2.23.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
